Add extensions.yaml (subset of RHCOS)
This is a subset of the `extensions.yaml` from
https://github.com/openshift/os

My immediate motivation is that day to day I develop in
a Fedora toolbox targeting FCOS, and I wanted to test
`rpm-ostree compose extensions` as `cosa buildextend-extensions`
invokes it, and it's easier if we have a file here.

Since the FCOS pipeline doesn't invoke `cosa buildextend-extensions`,
this will be a no-op.

Or stated more strongly: We might just use this as part of e.g.
rpm-ostree's CI runs.

But...one could imagine that we actually share the list of extensions
between FCOS and RHCOS in the future.

Or more obviously, perhaps we actually do something with this FCOS
extensions list and don't actually offer "all Fedora packages"
as extensions.